Crate toy_rsa_lib

Utility routines for toy RSA implementation.

This code provides some functions useful for implementing RSA with toy-sized (62-bit public) keys.

Functions

gcd

Compute the greatest common divisor of two positive numbers.

lcm

Compute the least common multiple of two positive numbers.

modexp

Efficiently compute x**y mod m. Uses the standard approach of repeated squaring with adjustments. O(lg y) runtime.

modinverse

Compute the modular inverse of a in the field mod m. m must be greater than 0 and coprime with a.

rsa_prime

Produce a prime in the range 2**31..=2**32-1 suitable for use in RSA.